test code for non-1st port usage
If a user doesn't use the 1st port- and only uses 2,3,4 etc, or 2,4,5 then there is an issue with old code as it will continue to use ports 1,2,3,4 etc.

This will cause no sound to generate as port 1 is important multiplication.  This addition is work in progress, at this point may work.